U-Turn design
As the aim of this project is to offer a base for creating rpg games, it
should not make any constraint.
In the previous version, a class derived from ICharacter had to define
methods "getBaseAttack" and "setDamages" (and more others would be
added)
BUT this forces designers to use theses data ("attack" and "damages") to
make their own game.
As a game designer I'd like to make my own attributes such as "attack",
"hp" and others but also "someSpecialAttrobuteForMyVerySpecialGame".
With the previous design, it would have been very horrible to add (one
should have to recode many parts of this library which is NOT
effective).

So I decided to change my mind and to propose attributes setter and
getter. An attribute is something such as "hp", or "damages" or whatever
a game designer wants !
This should be useful for ISkill which can now access to game-specific
data of a ICharacter to make their own effects.

This new design should allow user to define how a ISkill should effect a
ICharacter BUT as a game designer I would like to say "I want that this
skill will affect THIS attribute of the target by THIS equation.".
So I added a math expression parser which allows user to define his own
equations for his game (see uut.cpp for usage).

This first version of new design is not fully developped. Changes will
happen later.